Move numpy_utils to utils.numpy
